Create InverseDocumentFrequency.py
which measures how important a term is. While computing TF, all terms are considered equally important. However, it is known that certain terms, such as "is", "of", and "that", may appear a lot of times but have little importance. Thus we need to weigh down the frequent terms while scaling up the rare ones, by computing the following: 

IDF(t) = log_e(Total number of documents / Number of documents with term t in it).